Add Toolio and WorldLoom extensions for PDL processing
Add `ToolioCompletion` and `WorldLoomRead` as extensions (re-organizing how PDL constructs can be added) to enhance functionality with structured LLM responses. Updates dispatcher resolution order to include the new extensions, Moved `WorldLoomRead` implementation from `object_pdl_model.py` to a dedicated module.
